large change to test structure and coverage. These tests are mostly AI generated and will be manually reviewed as time permits
