+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "oshconnect"
-version = "0.5.0a0"
+version = "0.5.0a1"
 description = "Library for interfacing with OSH, helping guide visualization efforts, and providing a place to store configurations. Implements OGC CS API Part 3 (Pub/Sub) MQTT topic conventions including :data topics and resource event topics."
 readme = "README.md"
 authors = [
@@ -31,3 +31,6 @@ packages = {find = { where = ["src/"]}}
 
 [tool.pytest.ini_options]
 pythonpath = ["src"]
+markers = [
+    "network: test requires a live OSH server or external network endpoint (skipped by default in CI; see workflow `tests.yaml`).",
+]
